refactor(core): lift readMultipleFilesFromProjectDir to core; decouple ask/exec from review
Move readMultipleFilesFromProjectDir into @gaunt-sloth/core/utils/fileUtils and
delete its local definition in review. review's fileUtils still re-exports core
via `export *`, so the @gaunt-sloth/review/utils/fileUtils deep-import keeps
working for review's own consumers (reviewCommand, prCommand, etc.).

- ask/exec now import readMultipleFilesFromProjectDir from
  @gaunt-sloth/core/utils/fileUtils.js and import nothing from
  @gaunt-sloth/review; their specs mock the core path
- add real readMultipleFilesFromProjectDir tests to core's fileUtils.spec
