+++ packages/assistant/spec/askCommand.spec.ts
@@ -37,7 +37,7 @@ const fileUtilsMock = {
   fileSafeLocalDate: vi.fn(),
   generateStandardFileName: vi.fn(),
 };
-vi.mock('@gaunt-sloth/review/utils/fileUtils.js', () => fileUtilsMock);
+vi.mock('@gaunt-sloth/core/utils/fileUtils.js', () => fileUtilsMock);
 
 // Mock config to return specific test values
 const mockConfig = {
